+++ terraform/modules/rke1/data.tf
@@ -3,13 +3,6 @@ data "openstack_networking_network_v2" "ext_net" {
   name = var.openstack_external_net
 }
 
-# boot image
-# DEPRECATED
-data "openstack_images_image_v2" "boot" {
-  name        = var.os
-  most_recent = true
-}
-
 # openstack project name (bbXX)
 data "openstack_identity_auth_scope_v3" "scope" {
   name = "my_scope"
@@ -17,6 +10,6 @@ data "openstack_identity_auth_scope_v3" "scope" {
 
 data "openstack_images_image_v2" "os_image" {
   for_each    = var.openstack_os_image
-  name        = each.value
+  name        = each.value.imagename
   most_recent = true
 }